New approach is yielding some fruits. Cards are almost done. Tripped up a lot with the Vectors, but i think i could get this working within the next hour. Again, to reiterate, the game is currently broken up into halves: the Card portion of the game, and the Chess Portion of the game. Other features such as Economy and modular puzzles will most likely be glossed over in favor of static puzzles that are hard coded into an instance of a control panel. Control Panels contain game level information such as the chessboard, deck size, and gamepieces available to the player.
I'm probably going to shift Deck.cpp into a vector based format since the position of the cards could prove useful for reshuffling used cards back into the deck.
With the deque set up the way it is,the card that a player draws will be placed on the bottom. However, a possible work around is to record the first card after the beginning shuffle and have a function that reads each card as its drawn. Once the First Card is read, the deck can be shuffled.